API/requirements.txt
fastapi
strawberry-graphql
uvicorn
neo4j
flask